+++ src/sequentially.ts
@@ -1,38 +1,37 @@
 import { isPromise } from './is_promise';
-import { Tail, First, Length, HasPromise, ReturnTypesOf } from './utilities';
+import {
+  First,
+  HasPromise,
+  ReturnTypesOf,
+  IsParametersEqual,
+} from './utilities';
 
-type IsParametersEqual<
-  F extends ReadonlyArray<(...values: ReadonlyArray<any>) => unknown>,
-  R = []
-> = Length<F> extends 0
-  ? R extends false
-    ? false
-    : true
-  : IsParametersEqual<
-      Tail<F>,
-      R extends []
-        ? Parameters<First<F>>
-        : R extends Parameters<First<F>>
-        ? R
-        : false
-    >;
+const concat = <V, A extends ReadonlyArray<V>>(
+  aggregator: A | Promise<A>,
+  result: V
+) =>
+  isPromise<A>(aggregator)
+    ? aggregator.then((a) => a.concat(result))
+    : aggregator.concat(result);
 
 /**
  * Independently invokes functions with the same values
  * in order that they are passed. Can handle asynchronous
  * functions.
  */
-export const sequentially = <
-  V extends ReadonlyArray<(...values: ReadonlyArray<any>) => unknown>
->(
-  ...fns: V
-) => (
-  ...values: IsParametersEqual<V> extends true
-    ? Parameters<First<V>>
-    : ReadonlyArray<unknown>
-): HasPromise<ReturnTypesOf<V>> extends true ? Promise<void> : void =>
-  fns.reduce(
-    (result, fn) =>
-      isPromise(result) ? result.then(() => fn(...values)) : fn(...values),
-    undefined as any
-  );
+export const sequentially =
+  <V extends ReadonlyArray<(...values: ReadonlyArray<any>) => unknown>>(
+    ...fns: V
+  ) =>
+  (
+    ...values: IsParametersEqual<V> extends true ? Parameters<First<V>> : never
+  ): HasPromise<V> extends true
+    ? Promise<ReturnTypesOf<V>>
+    : ReturnTypesOf<V> =>
+    fns.reduce((aggregator, fn) => {
+      const result = fn(...values);
+
+      return isPromise(result)
+        ? result.then((value) => concat(aggregator, value))
+        : concat(aggregator, result);
+    }, [] as any);